# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "G&L BigCommerce API",
        "description": "G&L Integration API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/uploadFile": {
            "get": {
                "summary": "Test API",
                "description": "Sube el Archivo",
                "operationId": "a32c6f28ab027a50f952284a684f06cc",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "API test route works!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/testentrada": {
            "get": {
                "summary": "Test API entrada",
                "description": "Returns a test message",
                "operationId": "940dc183a442fdf87db1c9c150e4acbd",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "API test route works!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/process-excel": {
            "get": {
                "summary": "Test API",
                "description": "Procesa el Excel",
                "operationId": "cbf5aa0a947be233d64c22d5de935515",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "API test route works!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    }
}